+++ assets/javascripts/kitten/components/cards/crowdfunding-card/components/button.js
@@ -5,13 +5,19 @@ import { Button } from '../../../../components/buttons/button/button'
 import { pxToRem } from '../../../../helpers/utils/typography'
 import COLORS from '../../../../constants/colors-config'
 
+const COMPONENT_GUTTER = pxToRem(10)
+
 const StyledLoading = styled(Button)`
   background-color: ${COLORS.line2};
   border-color: ${COLORS.line2};
 `
 
+const StyledButtonContainer = styled.div`
+  padding: 0 ${COMPONENT_GUTTER};
+`
+
 const CardButton = ({ text, loading }) => (
-  <>
+  <StyledButtonContainer>
     {!loading && (
       <Button fluid modifier="helium">
         {text}
@@ -19,7 +25,7 @@ const CardButton = ({ text, loading }) => (
     )}
 
     {loading && <StyledLoading fluid />}
-  </>
+  </StyledButtonContainer>
 )
 
 export default CardButton