+++ assets/javascripts/kitten/components/cards/crowdfunding-card/components/title.js
@@ -19,6 +19,10 @@ const StyledTitle = styled.div`
   line-height: 1;
 `
 
+const StyledWidgetTitle = styled(StyledTitle)`
+  padding: 0;
+`
+
 const StyledTitleLoading = styled.span`
   display: block;
   background-color: ${COLORS.line2};
@@ -38,7 +42,7 @@ const StyledWidgetContainer = styled.div`
 
 const StyledDayCounter = styled.div`
   line-height: 1;
-  margin-left: calc(4 * ${COMPONENT_GUTTER});
+  margin-left: calc(2 * ${COMPONENT_GUTTER});
 `
 
 const StyledStateDay = styled(Text)`
@@ -128,7 +132,7 @@ class TitleComponent extends PureComponent {
     return (
       <>
         <StyledWidgetContainer>
-          <StyledTitle>
+          <StyledWidgetTitle>
             {!loading && (
               <Text
                 tag="span"
@@ -143,7 +147,7 @@ class TitleComponent extends PureComponent {
                 {!titleTruncate && widgetTitle}
               </Text>
             )}
-          </StyledTitle>
+          </StyledWidgetTitle>
 
           {!loading && (
             <StyledDayCounter>