+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `color` and `title` props on `CopyIcon` component.
+- Feature: Create `EmbedIcon` component.
+
 ## [2.35.0] - 2019-10-29
 
 Feature: